Producer/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colour==0.1.5
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==19.3.0
httplib2==0.17.1
idna==2.9
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kafka-python==2.0.1
MarkupSafe==1.1.1
openpyxl==3.0.3
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
